attestation: refactor certificate caching into separate package
The TDX validator will also require a mechanism to store certificates, and as we have such code already present for SNP, just use it for both.
