+++ internal/attestation/certcache/cached_client_test.go
@@ -1,7 +1,7 @@
 // Copyright 2024 Edgeless Systems GmbH
 // SPDX-License-Identifier: AGPL-3.0-only
 
-package snp
+package certcache
 
 import (
 	"log/slog"
